final/requirements.txt
openai==0.28.1
spotipy==2.23.0
python-dotenv
flask
flask_cors
requests
